refactor: modularize select
